seshat-node: Move the helper methods into a separate module.
